safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network/network-content-replacement-embed-expected.txt

CONSOLE MESSAGE: line 58: Done.
Tests NetworkResourcesData logic for embed content replacement.


http://127.0.0.1:8000/inspector/network/resources/resource.php?size=200
resource.content after requesting content: null
http://127.0.0.1:8000/inspector/network/resources/resource.php?size=100
resource.content after requesting content: ****************************************************************************************************
http://127.0.0.1:8000/inspector/network/resources/resource.php?size=201
resource.content after requesting content: null
http://127.0.0.1:8000/inspector/network/resources/resource.php?size=100
resource.content after requesting content: ****************************************************************************************************
